feat: remove unused user functionality
