+++ src/store/reducers/index.ts
@@ -5,7 +5,6 @@ import chessgroundReducer from 'store/reducers/chessgroundReducer';
 import gameReducer from 'store/reducers/gameReducer';
 import leaderboardReducer from 'store/reducers/leaderboardReducer';
 import requestReducer from 'store/reducers/requestReducer';
-import userReducer from 'store/reducers/userReducer';
 import wagerReducer from 'store/reducers/wagerReducer';
 
 const rootReducer = combineReducers({
@@ -14,7 +13,6 @@ const rootReducer = combineReducers({
   game: gameReducer,
   leaderboard: leaderboardReducer,
   requests: requestReducer,
-  user: userReducer,
   wager: wagerReducer,
 });
