+++ src/types/state.ts
@@ -9,7 +9,6 @@ import { SocketActions, SocketActionTypes } from 'types/socket';
 
 import { AuthActions, AuthActionTypes, AuthState } from 'types/resources/auth';
 import { GameActions, GameActionTypes, GameState } from 'types/resources/game';
-import { UserActions, UserActionTypes } from 'types/resources/user';
 import { WagerActions, WagerActionTypes, WagerState } from 'types/resources/wager';
 import { CgActions, CgActionTypes, ChessgroundState } from './chessground';
 import { LeaderboardActions, LeaderboardActionTypes, LeaderboardState } from './leaderboard';
@@ -18,12 +17,10 @@ import { LeaderboardActions, LeaderboardActionTypes, LeaderboardState } from './
 
 export type Actions =
   AuthActions | GameActions | SocketActions |
-  UserActions | WagerActions | CgActions |
-  LeaderboardActions;
+  WagerActions | CgActions | LeaderboardActions;
 export type ActionTypes =
   AuthActionTypes | GameActionTypes | SocketActionTypes |
-  UserActionTypes | WagerActionTypes | CgActionTypes |
-  LeaderboardActionTypes;
+  WagerActionTypes | CgActionTypes | LeaderboardActionTypes;
 
 export const REQUEST = 'REQUEST';
 export const SUCCESS = 'SUCCESS';
